Updated the skin implementation process
This version of the process is how you would do it for uPortal-Start. Under the assumption that unicon builds the binaries that the rest of us reference, uPortal-Start will be the way that everyone will be interacting with this in the future. So while it may not be completely accurate to the this particular repository, it is accurate for the majority of cases of people who will see this guide.